Switch to DataJpaTest with testcontainers; Refactor DAOs using JpaRepository; Add mysql 8.x testcontainer; Refactor tests using mysql and dbunit; Call test execution listeners in reversed order; User more Java 8+.

+++ rest/src/main/java/mybar/web/rest/rates/RatesController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -37,7 +38,7 @@ public class RatesController {
 
     @RequestMapping(value = "/average", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.OK)
-    public Map<String, Double> findAllAverageRates() {
+    public Map<String, BigDecimal> findAllAverageRates() {
         return ratesService.findAllAverageRates();
     }
